Add hydra:totalItems on partial pagination
When we’re on the last page, including the total # of items is useful
data and can be inferred.

This won’t work 100% of the time (occasionally item counts land exactly
on page boundaries, and so will still be omittted), but should still be
useful in the cases it does.